# #4314 [FIX] performance all of check budget