- แก้ไขให้ระบบ query `from budget_monitor_report`
แทน from แต่ละ module (res_project_monitor_view, res_section_monitor_view)
- แก้ไข code WA, Purchase Invoice Plan ให้เป็น best pattern

ผลการทดสอบ check budget
PR
15 -> 8 s

PO Confirm
6 -> 0.4 s

HR Expense
7.3 -> 4 s

INV
6 -> 0.4 s

IN Transfer -> Asset (10 assets)
93 -> 54 s

Deployment : restart only